Move hook away from Entity Analytics folder
